Improve getDefaultSecurityGroupId()
If there's no default VPC but there is only 1 VPC, consider it the
default. If there is no default security group in a VPC, but there is
only 1 security group, then also consider it the default. This enables
spotsh to succeed by default in AWS environments where the user had
deleted the default VPC or a VPC's default security group.