Pluralize --type launch argument
This commit replaces the singular --type launch flag with a comma
separated plural --types flag. As a consequence the instance type
preference is similarly pluralized. Any users upgrading should first
delete any prior preferences ($HOME/.config/spotsh/prefs.json) and
re-run 'spotsh config'.

The motivation behind this change is that when working with multiple
regions, often it is the case that an instance type that is available
in 1 region is not available in another. In the common case the user
doesn't care and just wants a cheap x86_64 linux instance, so to avoid
bothering the user by forcing a different --type flag per region we
just pluralized the instance type.