Automatically add ssh ingress rule if needed
When launching a spot instance in a region that the user has never
previously used, it is likely that the user has not configured the
default security group's ingress rules to allow for ssh
connections. This commit updates spotsh to automatically add an
ingress rule to the security group if needed. This also enables spotsh
to "just work" out of the box for completely new users to AWS.